requirements.txt
fastapi
uvicorn
torch
torchvision
pillow
requests
numpy
python-multipart